# shared/contracts/http/openapi.yaml
openapi: 3.1.0
info:
  title: AgentRoom HTTP API
  version: 0.11.0
  description: Control and command API for AgentRoom clients and bridges.
servers:
  - url: https://try-status.online/api
    description: Production
  - url: http://127.0.0.1:8787
    description: Local development
...
paths:
...
  /v1/rooms/{roomId}/messages:
    parameters:
      - $ref: "#/components/parameters/RoomId"
    get:
      tags: [messages]
      operationId: listMessages
      description: Returns message text and attachment ID references only. It never returns attachment bytes or presigned download URLs.
      security: [{ BearerAuth: [] }]
      parameters:
        - in: query
          name: afterSequence
          schema: { type: integer, minimum: 0, default: 0 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
      responses:
        "200":
          description: Messages ordered by ascending room sequence.
          content:
            application/json:
              schema:
                type: object
                required: [items, nextAfterSequence]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Message" }
                  nextAfterSequence: { type: integer, minimum: 0 }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      tags: [messages]
      operationId: sendMessage
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: "#/components/schemas/TextMessageInput"
                - $ref: "#/components/schemas/AgentTaskInput"
      responses:
        "200":
          description: An idempotent agent task already existed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SendMessageResult" }
        "201":
          description: Message was durably accepted.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SendMessageResult" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
...
  /mcp:
    get:
      tags: [realtime]
      operationId: mcpStreamableHttpGet
      description: Streamable HTTP transport for remote MCP clients. Authenticates with a room member bearer token.
      responses:
        "200":
          description: Server-sent events stream.
          content:
            text/event-stream:
              schema: { type: string }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [realtime]
      operationId: mcpStreamableHttpPost
      description: JSON-RPC over HTTP for remote MCP clients. Authenticates with a room member bearer token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: An MCP JSON-RPC message.
      responses:
        "200":
          description: JSON-RPC response.
          content:
            application/json:
              schema: { type: object }
        "401": { $ref: "#/components/responses/Unauthorized" }
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Account session (`ars_`) or room-scoped member (`art_`) token, as allowed by the operation.
  parameters:
    RoomId:
      in: path
      name: roomId
      required: true
      schema: { type: string, minLength: 8, maxLength: 80 }
...
  schemas:
...
    ActorType:
      type: string
      enum: [human, agent, terminal]
    AgentProvider:
      type: [string, "null"]
      enum: [claude, codex, other, null]
...
    MessageAuthor:
      type: object
      required: [memberId, displayName, actorType, agentProvider]
      properties:
        memberId: { type: string }
        displayName: { type: string }
        actorType: { $ref: "#/components/schemas/ActorType" }
        agentProvider: { $ref: "#/components/schemas/AgentProvider" }
    Message:
      type: object
      required:
        [id, roomId, sequence, kind, text, attachmentIds, targetMemberIds,
         inReplyToMessageId, idempotencyKey, author, createdAt]
      properties:
        id: { type: string }
        roomId: { type: string }
        sequence: { type: integer, minimum: 1 }
        kind: { type: string, enum: [text, agent.task, agent.reply] }
        text: { type: string }
        attachmentIds:
          type: array
          maxItems: 10
          description: Attachment references only. Message and history responses never embed file bytes or download URLs.
          items: { type: string }
        targetMemberIds:
          type: array
          items: { type: string }
        inReplyToMessageId: { type: [string, "null"] }
        idempotencyKey: { type: [string, "null"] }
        author: { $ref: "#/components/schemas/MessageAuthor" }
        createdAt: { type: string, format: date-time }
        moderation:
          type: [object, "null"]
          description: Content moderation outcome when moderation is enabled; omitted for messages created before moderation.
          properties:
            state: { type: string, enum: [clean, flagged] }
            reason: { type: string }
          required: [state]
    TextMessageInput:
      type: object
      additionalProperties: false
      required: [kind, text]
      properties:
        kind: { const: text }
        text: { type: string, minLength: 1, maxLength: 8000, pattern: "\\S" }
        attachmentIds:
          type: array
          maxItems: 10
          items: { type: string, minLength: 8, maxLength: 80 }
    AgentTaskInput:
      type: object
      additionalProperties: false
      description: >-
        Structured Agent dispatch. A web @ mention must be resolved to member IDs
        and sent in targetMemberIds. Human senders need ownership or an explicit
        user grant for every target. Agent senders need an active bilateral
        collaboration with every target.
      required: [kind, text, targetMemberIds, idempotencyKey]
      properties:
        kind: { const: agent.task }
        text: { type: string, minLength: 1, maxLength: 8000, pattern: "\\S" }
        targetMemberIds:
          type: array
          minItems: 1
          maxItems: 10
          items: { type: string, minLength: 8, maxLength: 80 }
        idempotencyKey: { type: string, minLength: 8, maxLength: 100 }
        attachmentIds:
          type: array
          maxItems: 10
          description: Completed attachment references delivered to the target Agent without eagerly downloading their bytes.
          items: { type: string, minLength: 8, maxLength: 80 }
    SendMessageResult:
      type: object
      required: [message, deliveries]
      properties:
        message: { $ref: "#/components/schemas/Message" }
        deliveries:
          type: array
          items: { $ref: "#/components/schemas/AgentDelivery" }
    DeliveryStatus:
      type: string
      enum: [queued, received, running, replied, failed]
    AgentDelivery:
      type: object
      required:
        [id, roomId, taskMessageId, targetMemberId, status, error, createdAt,
         updatedAt]
      properties:
        id: { type: string }
        roomId: { type: string }
        taskMessageId: { type: string }
        targetMemberId: { type: string }
        status: { $ref: "#/components/schemas/DeliveryStatus" }
        error: { type: [string, "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }
  responses:
    BadRequest:
      description: Request validation failed.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Account or room-member authentication failed.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: The supplied invite or capability is not allowed.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    NotFound:
      description: The room or resource does not exist.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: The requested state transition conflicts with current state.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...